Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -58,5 +58,5 @@ def get_average_age_for_season(data, season)
       end 
     end 
   end 
-  (ages / counter).to_f
+  (ages / counter).to_i
 end
